backend/requirements.txt
cryptography==42.0.7
Django==4.2
django-filter==23.1
django-templated-mail==1.1.1
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
py==1.11.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
django-rest-authtoken==2.1.4